Agenda: Add detailed agenda
This adds table support in a really awful way, but I'm not a CSS expert,
I only play one when supporting Hugo websites. It seems as if the
standard Hugo shortcode way [1] of supporting tables doesn't work with
the Tachyon CSS engine [2] which our Ananke theme [3] uses.

To make this work, I added a rawhtml shortcode which allows me to update
the CSS stylesheet right before the table to make it display properly.

[1] https://willschenk.com/articles/2020/styling_tables_with_hugo/
[2] https://tachyons.io
[3] https://github.com/theNewDynamic/gohugo-theme-ananke

Signed-off-by: Kyle Mestery <mestery@mestery.com>